// payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    feature?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Team {
  id: number;
  name: string;
  /**
   * URL path, e.g. john-doe -> /team-details/john-doe
   */
  slug: string;
  designation?: string | null;
  image?: (number | null) | Media;
  bio?: string | null;
  facebook?: string | null;
  twitter?: string | null;
  instagram?: string | null;
  dribble?: string | null;
  updatedAt: string;
  createdAt: string;
}